Add files via upload
This repository provides and interactive navigation menu bar consisting of Home , About , Services and Contact navigation links. 
Along with the provided instruction (creating an interactive resoponsive navigation menu bar) , I have also provided a functionality of smooth scrolling that is whenever a navigation link is clicked it takes the user to that section smoothly, and along with this there are two contact options one is number and another is mail , which is displayed in a block manner when hovered over.